requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==19.3.0             # via pytest
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
flake8==3.8.3             # via -r requirements.in
idna==2.10                # via requests
importlib-metadata==1.7.0  # via flake8, pluggy, pytest
iniconfig==1.0.1          # via pytest
isort==5.2.2              # via -r requirements.in
mccabe==0.6.1             # via flake8
more-itertools==8.4.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.782               # via -r requirements.in
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pytest==6.0.1             # via -r requirements.in
python-dotenv==0.14.0     # via -r requirements.in
requests==2.24.0          # via -r requirements.in
six==1.15.0               # via packaging
toml==0.10.1              # via pytest
typed-ast==1.4.1          # via mypy
typing-extensions==3.7.4.2  # via mypy
urllib3==1.25.10          # via requests
zipp==3.1.0               # via importlib-metadata
